requirements.txt
patool
mosspy
bs4
